requirements.txt
yt-dlp
fpdf2
python-dotenv
streamlit
pandas
